# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Termosalud API",
        "description": "API RESTful para la plataforma global multi-mercado de equipamiento médico estético",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/v1/languages": {
            "get": {
                "tags": [
                    "Languages"
                ],
                "summary": "Listar Language",
                "description": "Listar Language",
                "operationId": "listLanguage",
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Languages"
                ],
                "summary": "Crear Language",
                "description": "Crear Language",
                "operationId": "createLanguage",
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar productos",
                "description": "Obtiene la lista de todos los productos disponibles",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Búsqueda por nombre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Campo por el que ordenar",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Dirección del orden (asc, desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Límite de resultados",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Desplazamiento para paginación",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "01jf2x3y4z5a6b7c8d9e0f1g"
                                                    },
                                                    "name": {
                                                        "properties": {
                                                            "es": {
                                                                "type": "string",
                                                                "example": "Zionic Pro Max"
                                                            },
                                                            "en": {
                                                                "type": "string",
                                                                "example": "Zionic Pro Max"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "sku": {
                                                        "type": "string",
                                                        "example": "ZIONIC-PRO-MAX"
                                                    },
                                                    "is_featured": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "status": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear Product",
                "description": "Crear Product",
                "operationId": "createProduct",
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/treatments": {
            "get": {
                "tags": [
                    "Treatments"
                ],
                "summary": "Listar tratamientos",
                "operationId": "getTreatments",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Búsqueda por nombre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Campo por el que ordenar",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Dirección del orden (asc, desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Límite de resultados",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Desplazamiento para paginación",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tratamientos obtenida exitosamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Treatments"
                ],
                "summary": "Crear Treatment",
                "description": "Crear Treatment",
                "operationId": "createTreatment",
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/treatment-categories/{id}": {
            "put": {
                "tags": [
                    "Treatment Categories"
                ],
                "summary": "Actualizar TreatmentCategory",
                "description": "Actualizar TreatmentCategory",
                "operationId": "updateTreatmentCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de TreatmentCategory",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Treatment Categories"
                ],
                "summary": "Eliminar TreatmentCategory",
                "description": "Eliminar TreatmentCategory",
                "operationId": "deleteTreatmentCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de TreatmentCategory",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Éxito"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación mediante token JWT de Laravel Passport",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}